DIJI_AI_BERT/requirements.txt
requests
pyyaml
joblib
pandas
scikit-learn
transformers
torch